rootfs_deb.bbclass: Add sanity check
git-svn-id: https://svn.o-hand.com/repos/poky/trunk@991 311d38ba-8fff-0310-9ca6-ca027cbcb966

+++ meta/classes/rootfs_deb.bbclass
@@ -102,7 +102,9 @@ fakeroot rootfs_deb_do_rootfs () {
 
 	# Hacks to make dpkg/ipkg coexist for now
 	mv ${IMAGE_ROOTFS}/var/dpkg ${IMAGE_ROOTFS}/usr/
-	rmdir ${IMAGE_ROOTFS}/usr/dpkg/alternatives
+	if [ -e ${IMAGE_ROOTFS}/usr/dpkg/alternatives ]; then
+		rmdir ${IMAGE_ROOTFS}/usr/dpkg/alternatives
+	fi
 	ln -s /usr/lib/ipkg/alternatives ${IMAGE_ROOTFS}/usr/dpkg/alternatives
 	ln -s /usr/dpkg/info ${IMAGE_ROOTFS}/usr/lib/ipkg/info
 	ln -s /usr/dpkg/status ${IMAGE_ROOTFS}/usr/lib/ipkg/status
